[TMP] Fix broken python override in bazel_rules_hdl
Removes call to pip_install_deps() (that is hardcoded to use python39)
from init() of bazel_rules_hdl. We later call it ourselves in way that
uses proper interpreter (python310).

The patch could probably be reverted in the future if we update hdl rules
to revision that integrates this commit:
https://github.com/hdl/bazel_rules_hdl/commit/7a1ba0e8d229200b4628e8a676917fc6b8e165d1