Clean up ram_mux
We apply implicit conditional I/O (inheritance of I/O predicates from branches) to simplify the logic of ram_mux. In addition, we note that in the current mux logic, we only ever have pending requests for the active channel, so we eliminate an unnecessary state element.

While we're here, we also switch to using more explicit names.